# schemas/yaml/bundled-openapi.yaml
openapi: 3.0.0
info:
  title: Event Service
  version: 1.0.0
servers:
  - url: http://localhost:8081
paths:
...
  /status:
    get:
      tags:
        - Status
      summary: Получить все статусы
      responses:
        '200':
          description: Список всех статусов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StatusResponse'
    post:
      tags:
        - Status
      summary: Создать статус
      requestBody:
        $ref: '#/components/requestBodies/Status'
      responses:
        '200':
          description: Создание статуса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  /status/{Id}:
    get:
      tags:
        - Status
      summary: Получить статус по Id
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Полученный статус
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
    put:
      tags:
        - Status
      summary: Изменить статус по id
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        $ref: '#/components/requestBodies/StatusUpdate'
      responses:
        '200':
          description: Измененный статус
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
    delete:
      tags:
        - Status
      summary: Удалить статус по id
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Удаленная локация
...
  /timeline:
    get:
      tags:
        - Timeline
      summary: Получить таймлайн поинты по треку и статусу
      parameters:
        - $ref: '#/components/parameters/XTrackId'
        - $ref: '#/components/parameters/Status'
      responses:
        '200':
          description: Список таймлайн поинтов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimelineResponse'
    post:
      tags:
        - Timeline
      summary: Создать таймлайн поинт
      requestBody:
        $ref: '#/components/requestBodies/Timeline'
      responses:
        '200':
          description: Созданный таймлайн
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineResponse'
...
  /timeline/{Id}:
    get:
      tags:
        - Timeline
      summary: Получить таймлайн поинт по Id
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Конкретный таймлайн поинт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineStatusResponse'
    put:
      tags:
        - Timeline
      summary: Изменить таймлайн поинт по Id
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        $ref: '#/components/requestBodies/TimelineUpdate'
      responses:
        '200':
          description: Обновленный таймлайн поинт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineStatusResponse'
    delete:
      tags:
        - Timeline
      summary: Изменить таймлайн поинт по Id
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Удаленный таймлайн поинт
components:
  schemas:
...
    Status:
      required:
        - title
      type: object
      properties:
        title:
          type: string
          example: Status Title
    StatusResponse:
      required:
        - id
      allOf:
        - $ref: '#/components/schemas/Status'
        - type: object
          properties:
            id:
              type: integer
              example: 42
    StatusUpdate:
      type: object
      properties:
        title:
          type: string
          example: Status title
...
    Timeline:
      required:
        - title
        - description
        - deadline
        - is_blocking
        - status
        - track_id
        - timeline_status_id
      type: object
      properties:
        title:
          type: string
          example: Timeline
        description:
          type: string
          example: Description
        deadline:
          type: string
          format: date-time
          example: '2020-01-01T00:00:00Z'
        is_blocking:
          type: boolean
          example: true
        status:
          type: string
          example: ready
        track_id:
          type: integer
          example: 1
        timeline_status_id:
          type: integer
          example: 1
    TimelineResponse:
      required:
        - id
      allOf:
        - $ref: '#/components/schemas/Timeline'
        - type: object
          properties:
            id:
              type: integer
              example: 42
    TimelineStatusResponse:
      required:
        - count_num
      properties:
        count_num:
          type: integer
          example: 1
    TimelineUpdate:
      type: object
      properties:
        title:
          type: string
          example: Timeline
        description:
          type: string
          example: Description
        deadline:
          type: string
          format: date-time
          example: '2020-01-01T00:00:00Z'
        is_blocking:
          type: string
          example: 'true'
        status:
          type: string
          example: ready
        track_id:
          type: integer
          example: 1
        timeline_status_id:
          type: integer
          example: 1
  requestBodies:
...
    Status:
      description: Данные для создания нового статуса
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
    StatusUpdate:
      description: Данные для обновления статуса
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StatusUpdate'
...
    Timeline:
      description: Данные для создания новой временной шкалы
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Timeline'
    TimelineUpdate:
      description: Данные для обновления временной шкалы
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TimelineUpdate'
  parameters:
    Id:
      name: Id
      in: path
      required: true
      description: Уникальный идентификатор
      schema:
        type: integer
...
    XTrackId:
      name: XTrackId
      in: header
      required: false
      description: Header идентификатор трека
      schema:
        type: integer
    Status:
      name: Status
      in: query
      required: false
      description: Статус
      schema:
        type: string